rf: move `JsonschemaValidationErrorType`
From `cli/models.py` to `models.py`
