+++ src/dandisets_linkml_status_tools/cli/models.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from typing import Annotated, Any, NamedTuple
+from typing import Annotated, Any
 
 from dandi.dandiapi import VersionStatus
 from linkml.validator.report import ValidationResult
@@ -67,26 +67,3 @@ class DandisetValidationReport(BaseModel):
 
     # Errors encountered in validation against the dandiset metadata model in LinkML
     linkml_validation_errs: LinkmlValidationErrsType = []
-
-
-class JsonschemaValidationErrorType(NamedTuple):
-    """
-    A named tuple for representing types of `jsonschema.exceptions.ValidationError`
-    objects.
-
-    The type of a `jsonschema.exceptions.ValidationError` is decided by the value of its
-    `validator` field and the value of its `validator_value` field. The values
-    of these fields are bundled in an instance of this named tuple to represent a type
-    of `jsonschema.exceptions.ValidationError` objects.
-    """
-
-    validator: str
-    validator_value: Any
-
-    def __eq__(self, other: object) -> bool:
-        return (
-            isinstance(other, JsonschemaValidationErrorType)
-            and self.validator == other.validator
-            and type(self.validator_value) is type(other.validator_value)  # noqa E721
-            and self.validator_value == other.validator_value
-        )
