+++ tests/test_cli/test_tools.py
@@ -2,7 +2,7 @@ import pytest
 from jsonschema.exceptions import ValidationError
 from linkml.validator.report import Severity, ValidationResult
 
-from dandisets_linkml_status_tools.cli.models import JsonschemaValidationErrorType
+from dandisets_linkml_status_tools.models import JsonschemaValidationErrorType
 from dandisets_linkml_status_tools.cli.tools import get_linkml_err_counts
